+++ src/avt/Blueprint/CMakeLists.txt
@@ -1,0 +1,30 @@
+# Copyright (c) Lawrence Livermore National Security, LLC and other VisIt
+# Project developers.  See the top-level LICENSE file for dates and other
+# details.  No copyright assignment is required to contribute to VisIt.
+
+#****************************************************************************
+#
+# Modifications:
+#
+#****************************************************************************
+
+set(AVTBLUEPRINT_SOURCES
+    avtConduitBlueprintDataAdaptor.C
+    avtConduitBlueprintInfoWarningHandler.C
+    )
+
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}
+                    ${VISIT_COMMON_INCLUDES}
+                    ${VTK_INCLUDE_DIRS}
+                    ${CONDUIT_INCLUDE_DIR}
+                    ${MFEM_INCLUDE_DIR}
+                    )
+
+# Add link directories
+link_directories(${LIBRARY_OUTPUT_DIRECTORY} ${CONDUIT_LIBRARY_DIR} )
+
+add_library(avtblueprint ${AVTBLUEPRINT_SOURCES})
+
+target_link_libraries(avtblueprint ${CONDUIT_LIB} vtkCommonCore vtkFiltersCore)
+
+VISIT_INSTALL_TARGETS(avtblueprint)
