+++ src/databases/Blueprint/CMakeLists.txt
@@ -18,7 +18,6 @@ SET(LIBM_SOURCES
 BlueprintMDServerPluginInfo.C
 ${COMMON_SOURCES}
 avtBlueprintFileFormat.C
-avtBlueprintDataAdaptor.C
 avtBlueprintTreeCache.C
 avtBlueprintOptions.C
 )
@@ -27,7 +26,6 @@ SET(LIBE_SOURCES
 BlueprintEnginePluginInfo.C
 ${COMMON_SOURCES}
 avtBlueprintFileFormat.C
-avtBlueprintDataAdaptor.C
 avtBlueprintTreeCache.C
 avtBlueprintWriter.C
 avtBlueprintOptions.C
@@ -36,12 +34,12 @@ avtBlueprintOptions.C
 INCLUDE_DIRECTORIES(
 ${CMAKE_CURRENT_SOURCE_DIR}
 ${CONDUIT_INCLUDE_DIR}
+${VISIT_INCLUDE_DIR}/avt/MFEM
+${VISIT_INCLUDE_DIR}/avt/Blueprint
 ${MFEM_INCLUDE_DIR}
 ${HDF5_INCLUDE_DIR}
 ${VISIT_DATABASE_INCLUDES}
 )
-
-ADD_DEFINITIONS(-DCONDUIT_HAVE_PARTITION_FLATTEN=${CONDUIT_HAVE_PARTITION_FLATTEN})
 
 LINK_DIRECTORIES(${VISIT_LIBRARY_DIR} ${MFEM_LIBRARY_DIR} ${CONDUIT_LIBRARY_DIR} )
 
@@ -51,19 +49,19 @@ SET(INSTALLTARGETS IBlueprintDatabase)
 
 IF(NOT VISIT_ENGINE_ONLY AND NOT VISIT_DBIO_ONLY)
     ADD_LIBRARY(MBlueprintDatabase ${LIBM_SOURCES})
-    TARGET_LINK_LIBRARIES(MBlueprintDatabase visitcommon avtdbatts avtdatabase_ser ${CONDUIT_LIB} ${MFEM_LIB} )
+    TARGET_LINK_LIBRARIES(MBlueprintDatabase visitcommon avtdbatts avtdatabase_ser avtmfem avtblueprint ${CONDUIT_LIB} ${MFEM_LIB} )
     ADD_TARGET_DEFINITIONS(MBlueprintDatabase MDSERVER)
     SET(INSTALLTARGETS ${INSTALLTARGETS} MBlueprintDatabase)
 ENDIF(NOT VISIT_ENGINE_ONLY AND NOT VISIT_DBIO_ONLY)
 
 ADD_LIBRARY(EBlueprintDatabase_ser ${LIBE_SOURCES})
-TARGET_LINK_LIBRARIES(EBlueprintDatabase_ser visitcommon avtpipeline_ser avtdatabase_ser ${CONDUIT_LIB} ${MFEM_LIB} )
+TARGET_LINK_LIBRARIES(EBlueprintDatabase_ser visitcommon avtpipeline_ser avtdatabase_ser avtmfem avtblueprint ${CONDUIT_LIB} ${MFEM_LIB} )
 ADD_TARGET_DEFINITIONS(EBlueprintDatabase_ser ENGINE)
 SET(INSTALLTARGETS ${INSTALLTARGETS} EBlueprintDatabase_ser)
 
 IF(VISIT_PARALLEL)
     ADD_PARALLEL_LIBRARY(EBlueprintDatabase_par ${LIBE_SOURCES})
-    TARGET_LINK_LIBRARIES(EBlueprintDatabase_par visitcommon avtpipeline_par avtdatabase_par ${CONDUIT_LIB} ${MFEM_LIB} ${CONDUIT_MPI_LIB} )
+    TARGET_LINK_LIBRARIES(EBlueprintDatabase_par visitcommon avtpipeline_par avtdatabase_par avtmfem avtblueprint ${CONDUIT_LIB} ${MFEM_LIB} ${CONDUIT_MPI_LIB} )
     ADD_TARGET_DEFINITIONS(EBlueprintDatabase_par ENGINE)
     SET(INSTALLTARGETS ${INSTALLTARGETS} EBlueprintDatabase_par)
 ENDIF(VISIT_PARALLEL)
