+++ src/databases/Blueprint/avtBlueprintFileFormat.C
@@ -60,7 +60,6 @@
 //-----------------------------------------------------------------------------
 #include "avtBlueprintLogging.h"
 #include "avtBlueprintTreeCache.h"
-#include "avtBlueprintDataAdaptor.h"
 
 #ifdef _WIN32
 #define strcasecmp stricmp
@@ -69,60 +68,6 @@
 using std::string;
 using namespace conduit;
 using namespace mfem;
-
-
-//-----------------------------------------------------------------------------
-// These methods are used to re-wire conduit's default error handling
-//-----------------------------------------------------------------------------
-void
-blueprint_plugin_print_msg(const std::string &msg,
-                           const std::string &file,
-                           int line)
-{
-    // Uncomment for very verbose traces:
-    //
-    // debug5 << "File:"    << file << std::endl
-    //        << "Line:"    << line << std::endl
-    //        << "Message:" << msg  << std::endl;
-    debug5 << msg  << std::endl;
-}
-
-//-----------------------------------------------------------------------------
-void
-blueprint_plugin_info_handler(const std::string &msg,
-                              const std::string &file,
-                              int line)
-{
-    blueprint_plugin_print_msg(msg,file,line);
-}
-
-
-//-----------------------------------------------------------------------------
-void
-blueprint_plugin_warning_handler(const std::string &msg,
-                                 const std::string &file,
-                                 int line)
-{
-    blueprint_plugin_print_msg(msg,file,line);
-}
-
-//-----------------------------------------------------------------------------
-void
-blueprint_plugin_error_handler(const std::string &msg,
-                               const std::string &file,
-                               int line)
-{
-    std::ostringstream bp_err_oss;
-    bp_err_oss << "[ERROR]"
-               << "File:"    << file << std::endl
-               << "Line:"    << line << std::endl
-               << "Message:" << msg  << std::endl;
-
-    debug1 << bp_err_oss.str();
-
-    BP_PLUGIN_EXCEPTION1(InvalidVariableException, bp_err_oss.str());
-
-}
 
 // ****************************************************************************
 //  Method: avtBlueprintFileFormat::FetchMeshAndTopoNames
@@ -177,6 +122,9 @@ sanitize_var_name(const std::string &varname)
 //    Added query to check MFEM LOR Setting. If it is set to legacy, the 
 //    constructor sets `m_new_refine` to false, and if it is set to new,
 //    the opposite. It is set to new by default.
+// 
+//    Justin Privitera, Wed Aug 24 11:18:25 PDT 2022
+//    Removed setting of info, warning, and error handlers.
 //
 // ****************************************************************************
 avtBlueprintFileFormat::avtBlueprintFileFormat(const char *filename, DBOptionsAttributes *opts)
@@ -198,13 +146,6 @@ avtBlueprintFileFormat::avtBlueprintFileFormat(const char *filename, DBOptionsAt
     }    
 
     m_tree_cache = new avtBlueprintTreeCache();
-
-    // these redirect conduit info and warnings to debug 5
-    conduit::utils::set_info_handler(blueprint_plugin_info_handler);
-    conduit::utils::set_warning_handler(blueprint_plugin_warning_handler);
-    // this catches any uncaught conduit errors, logs them to debug 1
-    // and  converts them into a VisIt Exception
-    conduit::utils::set_error_handler(blueprint_plugin_error_handler);
 }
 
 avtBlueprintFileFormat::~avtBlueprintFileFormat()
@@ -224,12 +165,27 @@ avtBlueprintFileFormat::~avtBlueprintFileFormat()
 //
 //  Programmer: harrison37 -- generated by xml2avt
 //  Creation:   Wed Jun 15 16:25:28 PST 2016
+// 
+//  Modifications:
+//     Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//     Encased in try-catch block.
 //
 // ****************************************************************************
 void
 avtBlueprintFileFormat::FreeUpResources(void)
 {
-    m_tree_cache->Release();
+    try
+    {
+        m_tree_cache->Release();
+    }
+    catch (conduit::Error &e)
+    {
+        std::ostringstream err_oss;
+        err_oss <<  "Conduit Exception in Blueprint Plugin "
+                    << "avtBlueprintFileFormat::FreeUpResources: " << endl
+                    << e.message();
+        BP_PLUGIN_EXCEPTION1(VisItException, err_oss.str());
+    }
 }
 
 // ****************************************************************************
@@ -1445,25 +1401,40 @@ avtBlueprintFileFormat::PopulateDatabaseMetaData(avtDatabaseMetaData *md)
 //
 //  Programmer: Cyrus Harrison
 //  Creation:   Thu Aug 11 10:15:05 PDT 2016
+// 
+//  Modifications:
+//     Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//     Encased in try-catch block.
 //
 // ****************************************************************************
 int
 avtBlueprintFileFormat::GetCycle()
 {
-    // VisIt doesn't support diff times / cycles for meshes in STMD
-    // we loop over all meshes and return the first valid cycle
-    NodeConstIterator itr = m_root_node["blueprint_index"].children();
-
-    while(itr.has_next())
+    try
     {
-        const Node &mesh = itr.next();
-        if(mesh.has_path("state/cycle"))
-        {
-            return mesh["state/cycle"].to_int();
-        }
-    }
+        // VisIt doesn't support diff times / cycles for meshes in STMD
+        // we loop over all meshes and return the first valid cycle
+        NodeConstIterator itr = m_root_node["blueprint_index"].children();
 
-    return avtFileFormat::INVALID_CYCLE;
+        while(itr.has_next())
+        {
+            const Node &mesh = itr.next();
+            if(mesh.has_path("state/cycle"))
+            {
+                return mesh["state/cycle"].to_int();
+            }
+        }
+
+        return avtFileFormat::INVALID_CYCLE;
+    }
+    catch (conduit::Error &e)
+    {
+        std::ostringstream err_oss;
+        err_oss <<  "Conduit Exception in Blueprint Plugin "
+                    << "avtBlueprintFileFormat::GetCycle: " << endl
+                    << e.message();
+        BP_PLUGIN_EXCEPTION1(VisItException, err_oss.str());
+    }
 }
 
 
@@ -1475,26 +1446,41 @@ avtBlueprintFileFormat::GetCycle()
 //
 //  Programmer: Cyrus Harrison
 //  Creation:   Thu Aug 11 10:15:05 PDT 2016
+// 
+//  Modifications:
+//     Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//     Encased in try-catch block.
 //
 // ****************************************************************************
 double
 avtBlueprintFileFormat::GetTime()
 {
-    // VisIt doesn't support diff times / cycles for meshes in STMD
-    // we loop over all meshes and return the first valid time
-
-    NodeConstIterator itr = m_root_node["blueprint_index"].children();
-
-    while(itr.has_next())
+    try
     {
-        const Node &mesh = itr.next();
-        if(mesh.has_path("state/time"))
-        {
-            return mesh["state/time"].to_double();
-        }
-    }
+        // VisIt doesn't support diff times / cycles for meshes in STMD
+        // we loop over all meshes and return the first valid time
 
-    return avtFileFormat::INVALID_TIME;
+        NodeConstIterator itr = m_root_node["blueprint_index"].children();
+
+        while(itr.has_next())
+        {
+            const Node &mesh = itr.next();
+            if(mesh.has_path("state/time"))
+            {
+                return mesh["state/time"].to_double();
+            }
+        }
+
+        return avtFileFormat::INVALID_TIME;
+    }
+    catch (conduit::Error &e)
+    {
+        std::ostringstream err_oss;
+        err_oss <<  "Conduit Exception in Blueprint Plugin "
+                    << "avtBlueprintFileFormat::GetTime: " << endl
+                    << e.message();
+        BP_PLUGIN_EXCEPTION1(VisItException, err_oss.str());
+    }
 }
 
 
@@ -1526,85 +1512,100 @@ avtBlueprintFileFormat::GetTime()
 //    Justin Privitera, Mon Apr 11 18:20:19 PDT 2022
 //    `m_new_refine` is passed to `RefineMeshToVTK` so it can choose the 
 //    appropriate MFEM LOR scheme.
+// 
+//     Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//     Encased in try-catch block.
 //
 // ****************************************************************************
 vtkDataSet *
 avtBlueprintFileFormat::GetMesh(int domain, const char *abs_meshname)
 {
-    BP_PLUGIN_INFO("GetMesh: "
-                   << abs_meshname << " [domain " << domain << "]");
-
-    // read mesh data into conduit tree
-    Node data;
-    string abs_meshname_str(abs_meshname);
-
-    // reads a single mesh into a blueprint conforming output
-    ReadBlueprintMesh(domain, abs_meshname_str, data);
-    // check for empty mesh case
-    if(data.dtype().is_empty())
+    try
     {
-        // support empty mesh case by returning NULL
-        return NULL;
-    }
+        BP_PLUGIN_INFO("GetMesh: "
+                       << abs_meshname << " [domain " << domain << "]");
 
-    Node verify_info;
-    if(!blueprint::mesh::verify(data, verify_info))
+        // read mesh data into conduit tree
+        Node data;
+        string abs_meshname_str(abs_meshname);
+
+        // reads a single mesh into a blueprint conforming output
+        ReadBlueprintMesh(domain, abs_meshname_str, data);
+        // check for empty mesh case
+        if(data.dtype().is_empty())
+        {
+            // support empty mesh case by returning NULL
+            return NULL;
+        }
+
+        Node verify_info;
+        if(!blueprint::mesh::verify(data, verify_info))
+        {
+            BP_PLUGIN_INFO("blueprint::mesh::verify failed for mesh "
+                           << abs_meshname << " [domain " << domain << "]" << endl
+                           << "Verify Info " << endl
+                           << verify_info.to_yaml() << endl
+                           << "Data Schema " << endl
+                           << data.schema().to_yaml());
+
+            BP_PLUGIN_INFO("warning: "
+                           "avtBlueprintFileFormat::GetMesh returning NULL "
+                           << abs_meshname
+                           << " [domain " << domain << "]"
+                           << " will be missing" << endl);
+            // TODO: Should we throw an error instead of blanking the domain?
+            return NULL;
+        }
+
+        BP_PLUGIN_INFO(data.schema().to_yaml());
+
+        // prepare result vtk dataset
+        vtkDataSet *res = NULL;
+
+
+        string mesh_name;
+        string topo_name;
+
+        FetchMeshAndTopoNames(std::string(abs_meshname),
+                              mesh_name,
+                              topo_name);
+
+        BP_PLUGIN_INFO("mesh name: " << mesh_name);
+        BP_PLUGIN_INFO("topo name: " << topo_name);
+
+        // check for the mfem case
+        if( m_mfem_mesh_map[topo_name] )
+        {
+            BP_PLUGIN_INFO("mesh  " << topo_name << " is a mfem mesh");
+            // use mfem to refine and create a vtk dataset
+            mfem::Mesh *mesh = avtConduitBlueprintDataAdaptor::BlueprintToMFEM::MeshToMFEM(data);
+            res = avtMFEMDataAdaptor::RefineMeshToVTK(mesh,
+                                                      domain,
+                                                      m_selected_lod+1,
+                                                      m_new_refine);
+
+            // cleanup the mfem mesh
+            delete mesh;
+        }
+        else
+        {
+            BP_PLUGIN_INFO("mesh  " << topo_name << " is a standard mesh");
+            // construct a vtk dataset directly from blueprint data
+            // in a conduit tree
+            res = avtConduitBlueprintDataAdaptor::BlueprintToVTK::MeshToVTK(domain, data);
+        }
+
+        BP_PLUGIN_INFO("avtBlueprintFileFormat::GetMesh Done");
+        return res;
+    }
+    catch (conduit::Error &e)
     {
-        BP_PLUGIN_INFO("blueprint::mesh::verify failed for mesh "
-                       << abs_meshname << " [domain " << domain << "]" << endl
-                       << "Verify Info " << endl
-                       << verify_info.to_yaml() << endl
-                       << "Data Schema " << endl
-                       << data.schema().to_yaml());
-
-        BP_PLUGIN_INFO("warning: "
-                       "avtBlueprintFileFormat::GetMesh returning NULL "
-                       << abs_meshname
-                       << " [domain " << domain << "]"
-                       << " will be missing" << endl);
-        // TODO: Should we throw an error instead of blanking the domain?
-        return NULL;
+        std::ostringstream err_oss;
+        err_oss <<  "Conduit Exception in Blueprint Plugin "
+                    << "avtBlueprintFileFormat::GetMesh: " << endl
+                    << e.message();
+        BP_PLUGIN_EXCEPTION1(VisItException, err_oss.str());
     }
-
-    BP_PLUGIN_INFO(data.schema().to_yaml());
-
-    // prepare result vtk dataset
-    vtkDataSet *res = NULL;
-
-
-    string mesh_name;
-    string topo_name;
-
-    FetchMeshAndTopoNames(std::string(abs_meshname),
-                          mesh_name,
-                          topo_name);
-
-    BP_PLUGIN_INFO("mesh name: " << mesh_name);
-    BP_PLUGIN_INFO("topo name: " << topo_name);
-
-    // check for the mfem case
-    if( m_mfem_mesh_map[topo_name] )
-    {
-        BP_PLUGIN_INFO("mesh  " << topo_name << " is a mfem mesh");
-        // use mfem to refine and create a vtk dataset
-        mfem::Mesh *mesh = avtBlueprintDataAdaptor::MFEM::MeshToMFEM(data);
-        res = avtBlueprintDataAdaptor::MFEM::RefineMeshToVTK(mesh, 
-                                                             m_selected_lod+1,
-                                                             m_new_refine);
-
-        // cleanup the mfem mesh
-        delete mesh;
-    }
-    else
-    {
-        BP_PLUGIN_INFO("mesh  " << topo_name << " is a standard mesh");
-        // construct a vtk dataset directly from blueprint data
-        // in a conduit tree
-        res = avtBlueprintDataAdaptor::VTK::MeshToVTK(domain, data);
-    }
-
-    BP_PLUGIN_INFO("avtBlueprintFileFormat::GetMesh Done");
-    return res;
 }
 
 // ****************************************************************************
@@ -1639,131 +1640,138 @@ avtBlueprintFileFormat::GetMesh(int domain, const char *abs_meshname)
 //    `m_new_refine` is passed to `RefineGridFunctionToVTK` so it can choose
 //    the appropriate MFEM LOR scheme.
 // 
+//     Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//     Encased in try-catch block.
+// 
+//    Justin Privitera, Tue Aug 23 14:40:24 PDT 2022
+//    Removed `CONDUIT_HAVE_PARTITION_FLATTEN` check.
+// 
 // ****************************************************************************
 
 vtkDataArray *
 avtBlueprintFileFormat::GetVar(int domain, const char *abs_varname)
 {
-    BP_PLUGIN_INFO("GetVar: " << abs_varname << " [domain " << domain << "]");
-
-    vtkDataArray *res = NULL;
-
-    string abs_varname_str(abs_varname);
-
-    // check for special var "element_coloring"
-    if(abs_varname_str.find("element_coloring") != std::string::npos)
+    try
     {
-        // TODO: we currently have to replace colons, brackets, etc
-        // to fetch from the mesh metadata, is there a standard helper
-        // util in VisIt that can take care of this for us?
-        string abs_meshname = metadata->MeshForVar(sanitize_var_name(abs_varname_str));
+        BP_PLUGIN_INFO("GetVar: " << abs_varname << " [domain " << domain << "]");
 
-        // element coloring is generated from the mesh
-        Node n_mesh;
-        // read the mesh data
-        ReadBlueprintMesh(domain, abs_meshname, n_mesh);
-        // check for empty mesh case
-        if(n_mesh.dtype().is_empty())
+        vtkDataArray *res = NULL;
+
+        string abs_varname_str(abs_varname);
+
+        // check for special var "element_coloring"
+        if(abs_varname_str.find("element_coloring") != std::string::npos)
         {
-            // support empty mesh case by returning NULL
-            return NULL;
-        }
+            // TODO: we currently have to replace colons, brackets, etc
+            // to fetch from the mesh metadata, is there a standard helper
+            // util in VisIt that can take care of this for us?
+            string abs_meshname = metadata->MeshForVar(sanitize_var_name(abs_varname_str));
 
-        Node verify_info;
-        if(!blueprint::mesh::verify(n_mesh,verify_info))
-        {
-            BP_PLUGIN_INFO("blueprint::mesh::verify failed for field "
-                           << abs_meshname << " [domain " << domain << "]" << endl
-                           << "Verify Info " << endl
-                           << verify_info.to_yaml() << endl
-                           << "Data Schema " << endl
-                           << n_mesh.schema().to_yaml());
-            return NULL;
-        }
-
-        // create an mfem mesh
-        mfem::Mesh *mesh = avtBlueprintDataAdaptor::MFEM::MeshToMFEM(n_mesh);
-        // refine the coloring to a vtk data array
-        res = avtBlueprintDataAdaptor::MFEM::RefineElementColoringToVTK(mesh,
-                                                                        domain,
-                                                                        m_selected_lod+1);
-        // clean up the mfem mesh
-        delete mesh;
-
-        // return the coloring result
-        return res;
-    }
-
-    // else, normal field case
-
-    Node n_field;
-    Node side_mesh;
-    Node *field_ptr = &n_field;
-
-    ReadBlueprintField(domain,abs_varname_str,*field_ptr);
-    // check for empty field case
-    if(field_ptr->dtype().is_empty())
-    {
-        // support empty field case by returning NULL
-        return NULL;
-    }
-
-    Node verify_info;
-    if(!blueprint::mesh::field::verify(*field_ptr,verify_info))
-    {
-        BP_PLUGIN_INFO("blueprint::mesh::field::verify failed for field "
-                       << abs_varname_str << " [domain " << domain << "]" << endl
-                       << "Verify Info " << endl
-                       << verify_info.to_yaml() << endl
-                       << "Data Schema " << endl
-                       << field_ptr->schema().to_yaml());
-        return NULL;
-    }
-
-    // if we have an association, this is a standard field
-    if(field_ptr->has_child("association"))
-    {
-        string abs_meshname = metadata->MeshForVar(sanitize_var_name(abs_varname_str));
-        // element coloring is generated from the mesh
-        Node n_mesh;
-        // read the mesh data
-        ReadBlueprintMesh(domain, abs_meshname, n_mesh);
-        // check for empty mesh case
-        if(n_mesh.dtype().is_empty())
-        {
-            // support empty mesh case by returning NULL
-            return NULL;
-        }
-
-        Node verify_info;
-        if(!blueprint::mesh::verify(n_mesh,verify_info))
-        {
-            BP_PLUGIN_INFO("blueprint::mesh::verify failed for field "
-                           << abs_meshname << " [domain " << domain << "]" << endl
-                           << "Verify Info " << endl
-                           << verify_info.to_yaml() << endl
-                           << "Data Schema " << endl
-                           << n_mesh.schema().to_yaml());
-            return NULL;
-        }
-
-        string mesh_name, topo_name;
-
-        FetchMeshAndTopoNames(std::string(abs_meshname),
-                              mesh_name,
-                              topo_name);
-
-        BP_PLUGIN_INFO("mesh name: " << mesh_name);
-        BP_PLUGIN_INFO("topo name: " << topo_name);
-
-        const Node &n_topo = n_mesh["topologies/" + topo_name];
-
-        if (n_topo.has_path("elements/shape"))
-        {
-            if (n_topo["elements/shape"].as_string() == "polyhedral" || 
-                n_topo["elements/shape"].as_string() == "polygonal")
+            // element coloring is generated from the mesh
+            Node n_mesh;
+            // read the mesh data
+            ReadBlueprintMesh(domain, abs_meshname, n_mesh);
+            // check for empty mesh case
+            if(n_mesh.dtype().is_empty())
             {
-                #if CONDUIT_HAVE_PARTITION_FLATTEN == 1
+                // support empty mesh case by returning NULL
+                return NULL;
+            }
+
+            Node verify_info;
+            if(!blueprint::mesh::verify(n_mesh,verify_info))
+            {
+                BP_PLUGIN_INFO("blueprint::mesh::verify failed for field "
+                               << abs_meshname << " [domain " << domain << "]" << endl
+                               << "Verify Info " << endl
+                               << verify_info.to_yaml() << endl
+                               << "Data Schema " << endl
+                               << n_mesh.schema().to_yaml());
+                return NULL;
+            }
+
+            // create an mfem mesh
+            mfem::Mesh *mesh = avtConduitBlueprintDataAdaptor::BlueprintToMFEM::MeshToMFEM(n_mesh);
+            // refine the coloring to a vtk data array
+            res = avtMFEMDataAdaptor::RefineElementColoringToVTK(mesh,
+                                                                 domain,
+                                                                 m_selected_lod+1);
+            // clean up the mfem mesh
+            delete mesh;
+
+            // return the coloring result
+            return res;
+        }
+
+        // else, normal field case
+
+        Node n_field;
+        Node side_mesh;
+        Node *field_ptr = &n_field;
+
+        ReadBlueprintField(domain,abs_varname_str,*field_ptr);
+        // check for empty field case
+        if(field_ptr->dtype().is_empty())
+        {
+            // support empty field case by returning NULL
+            return NULL;
+        }
+
+        Node verify_info;
+        if(!blueprint::mesh::field::verify(*field_ptr,verify_info))
+        {
+            BP_PLUGIN_INFO("blueprint::mesh::field::verify failed for field "
+                           << abs_varname_str << " [domain " << domain << "]" << endl
+                           << "Verify Info " << endl
+                           << verify_info.to_yaml() << endl
+                           << "Data Schema " << endl
+                           << field_ptr->schema().to_yaml());
+            return NULL;
+        }
+
+        // if we have an association, this is a standard field
+        if(field_ptr->has_child("association"))
+        {
+            string abs_meshname = metadata->MeshForVar(sanitize_var_name(abs_varname_str));
+            // element coloring is generated from the mesh
+            Node n_mesh;
+            // read the mesh data
+            ReadBlueprintMesh(domain, abs_meshname, n_mesh);
+            // check for empty mesh case
+            if(n_mesh.dtype().is_empty())
+            {
+                // support empty mesh case by returning NULL
+                return NULL;
+            }
+
+            Node verify_info;
+            if(!blueprint::mesh::verify(n_mesh,verify_info))
+            {
+                BP_PLUGIN_INFO("blueprint::mesh::verify failed for field "
+                               << abs_meshname << " [domain " << domain << "]" << endl
+                               << "Verify Info " << endl
+                               << verify_info.to_yaml() << endl
+                               << "Data Schema " << endl
+                               << n_mesh.schema().to_yaml());
+                return NULL;
+            }
+
+            string mesh_name, topo_name;
+
+            FetchMeshAndTopoNames(std::string(abs_meshname),
+                                  mesh_name,
+                                  topo_name);
+
+            BP_PLUGIN_INFO("mesh name: " << mesh_name);
+            BP_PLUGIN_INFO("topo name: " << topo_name);
+
+            const Node &n_topo = n_mesh["topologies/" + topo_name];
+
+            if (n_topo.has_path("elements/shape"))
+            {
+                if (n_topo["elements/shape"].as_string() == "polyhedral" || 
+                    n_topo["elements/shape"].as_string() == "polygonal")
+                {
                     string field_name, coords_name;
                     // get name of coordset from topology
                     coords_name = n_topo["coordset"].as_string();
@@ -1786,132 +1794,134 @@ avtBlueprintFileFormat::GetVar(int domain, const char *abs_varname)
                       options);
 
                     field_ptr = side_fields.fetch_ptr(field_name);
-                #else
-                    Node about;
-                    conduit::about(about);
-                    BP_PLUGIN_EXCEPTION1(InvalidVariableException,
-                        "VisIt Blueprint Plugin requires Conduit >= 0.8.0 to read polytopal meshes."
-                        "VisIt was built with Conduit version:"  << about["version"].as_string());
-                #endif
+                }
             }
+
+            // low-order case, use vtk
+            res = avtConduitBlueprintDataAdaptor::BlueprintToVTK::FieldToVTK(*field_ptr);
+        }
+        // if we have a basis, this field is actually an mfem grid function
+        else if(field_ptr->has_child("basis"))
+        {
+            // TODO: we currently have to replace colons, brackets, etc
+            // to fetch from the mesh metadata, is there a standard helper
+            // util in VisIt that can take care of this for us?
+            string abs_meshname = metadata->MeshForVar(sanitize_var_name(abs_varname_str));
+
+            // the grid function needs the mesh in order to refine
+
+            // read the mesh data
+            Node n_mesh;
+            ReadBlueprintMesh(domain, abs_meshname, n_mesh);
+            // check for empty mesh case
+            if(n_mesh.dtype().is_empty())
+            {
+                // support empty mesh case by returning NULL
+                return NULL;
+            }
+
+            Node verify_info;
+            if(!blueprint::mesh::verify(n_mesh,verify_info))
+            {
+                BP_PLUGIN_INFO("blueprint::mesh::verify failed for field "
+                               << abs_meshname << " [domain " << domain << "]" << endl
+                               << "Verify Info " << endl
+                               << verify_info.to_yaml() << endl
+                               << "Data Schema " << endl
+                               << n_mesh.schema().to_yaml());
+                return NULL;
+            }
+
+            // create an mfem mesh
+            mfem::Mesh *mesh = avtConduitBlueprintDataAdaptor::BlueprintToMFEM::MeshToMFEM(n_mesh);
+
+            // create the grid fuction
+            mfem::GridFunction *gf =  avtConduitBlueprintDataAdaptor::BlueprintToMFEM::FieldToMFEM(mesh,
+                                                                                 *field_ptr);
+            // refine the grid function into a vtk data array
+            res = avtMFEMDataAdaptor::RefineGridFunctionToVTK(mesh,
+                                                              gf,
+                                                              m_selected_lod+1,
+                                                              m_new_refine);
+
+            // cleanup mfem data
+            delete gf;
+            delete mesh;
         }
 
-        // low-order case, use vtk
-        res = avtBlueprintDataAdaptor::VTK::FieldToVTK(*field_ptr);
+        // check to see if we have matset_values multi-material data
+        // if so, provide them as AUXILIARY_DATA_MIXED_VARIABLE
+        if(field_ptr->has_child("matset"))
+        {
+            // get mesh and topo for var:
+            //  - abs_varname_str
+            
+            // replace colons, etc
+            abs_varname_str = sanitize_var_name(abs_varname_str);
+            string abs_meshname = metadata->MeshForVar(abs_varname_str);
+
+            BP_PLUGIN_INFO("field " << abs_varname << " is defined on mesh " << abs_meshname);
+
+            string mesh_name;
+            string topo_name;
+            FetchMeshAndTopoNames(std::string(abs_meshname),
+                                  mesh_name,
+                                  topo_name);
+
+            std::string matset_name = (*field_ptr)["matset"].as_string();
+            std::string mat_name = mesh_name + "_" + topo_name + "_" + matset_name;
+
+            BP_PLUGIN_INFO("mesh name: " << mesh_name);
+            BP_PLUGIN_INFO("topo name: " << topo_name);
+            BP_PLUGIN_INFO("materials name: " << mat_name);
+
+            // get the matset
+            // do the xform on the matset and the matset_values
+            Node n_matset;
+            ReadBlueprintMatset(domain,
+                                mat_name,
+                                n_matset);
+
+            Node n_silo_matset;
+            conduit::blueprint::mesh::field::to_silo(*field_ptr,
+                                                     n_matset,
+                                                     n_silo_matset);
+
+            int mix_len  = (int) n_silo_matset["field_mixvar_values"].dtype().number_of_elements();
+
+            float *mixvals_ptr = NULL;
+            if(n_silo_matset["field_mixvar_values"].dtype().is_float())
+            {
+                mixvals_ptr = n_silo_matset["field_mixvar_values"].as_float_ptr();
+            }
+            else
+            {
+                n_silo_matset["field_mixvar_values"].to_float_array(
+                                            n_silo_matset["field_mixvar_values_float"]);
+                mixvals_ptr = n_silo_matset["field_mixvar_values_float"].as_float_ptr();
+            }
+
+            avtMixedVariable *mvar = new avtMixedVariable(mixvals_ptr,
+                                                          mix_len,
+                                                          abs_varname_str);
+            void_ref_ptr mvar_ref = void_ref_ptr(mvar, avtMixedVariable::Destruct);
+            cache->CacheVoidRef(abs_varname_str.c_str(),
+                                AUXILIARY_DATA_MIXED_VARIABLE,
+                                timestep,
+                                domain,
+                                mvar_ref);
+        }
+
+        return res;
     }
-    // if we have a basis, this field is actually an mfem grid function
-    else if(field_ptr->has_child("basis"))
+    catch (conduit::Error &e)
     {
-        // TODO: we currently have to replace colons, brackets, etc
-        // to fetch from the mesh metadata, is there a standard helper
-        // util in VisIt that can take care of this for us?
-        string abs_meshname = metadata->MeshForVar(sanitize_var_name(abs_varname_str));
-
-        // the grid function needs the mesh in order to refine
-
-        // read the mesh data
-        Node n_mesh;
-        ReadBlueprintMesh(domain, abs_meshname, n_mesh);
-        // check for empty mesh case
-        if(n_mesh.dtype().is_empty())
-        {
-            // support empty mesh case by returning NULL
-            return NULL;
-        }
-
-        Node verify_info;
-        if(!blueprint::mesh::verify(n_mesh,verify_info))
-        {
-            BP_PLUGIN_INFO("blueprint::mesh::verify failed for field "
-                           << abs_meshname << " [domain " << domain << "]" << endl
-                           << "Verify Info " << endl
-                           << verify_info.to_yaml() << endl
-                           << "Data Schema " << endl
-                           << n_mesh.schema().to_yaml());
-            return NULL;
-        }
-
-        // create an mfem mesh
-        mfem::Mesh *mesh = avtBlueprintDataAdaptor::MFEM::MeshToMFEM(n_mesh);
-
-        // create the grid fuction
-        mfem::GridFunction *gf =  avtBlueprintDataAdaptor::MFEM::FieldToMFEM(mesh,
-                                                                             *field_ptr);
-        // refine the grid function into a vtk data array
-        res =  avtBlueprintDataAdaptor::MFEM::RefineGridFunctionToVTK(mesh,
-                                                                      gf,
-                                                                      m_selected_lod+1,
-                                                                      m_new_refine);
-
-        // cleanup mfem data
-        delete gf;
-        delete mesh;
-    }
-
-    // check to see if we have matset_values multi-material data
-    // if so, provide them as AUXILIARY_DATA_MIXED_VARIABLE
-    if(field_ptr->has_child("matset"))
-    {
-        // get mesh and topo for var:
-        //  - abs_varname_str
-        
-        // replace colons, etc
-        abs_varname_str = sanitize_var_name(abs_varname_str);
-        string abs_meshname = metadata->MeshForVar(abs_varname_str);
-
-        BP_PLUGIN_INFO("field " << abs_varname << " is defined on mesh " << abs_meshname);
-
-        string mesh_name;
-        string topo_name;
-        FetchMeshAndTopoNames(std::string(abs_meshname),
-                              mesh_name,
-                              topo_name);
-
-        std::string matset_name = (*field_ptr)["matset"].as_string();
-        std::string mat_name = mesh_name + "_" + topo_name + "_" + matset_name;
-
-        BP_PLUGIN_INFO("mesh name: " << mesh_name);
-        BP_PLUGIN_INFO("topo name: " << topo_name);
-        BP_PLUGIN_INFO("materials name: " << mat_name);
-
-        // get the matset
-        // do the xform on the matset and the matset_values
-        Node n_matset;
-        ReadBlueprintMatset(domain,
-                            mat_name,
-                            n_matset);
-
-        Node n_silo_matset;
-        conduit::blueprint::mesh::field::to_silo(*field_ptr,
-                                                 n_matset,
-                                                 n_silo_matset);
-
-        int mix_len  = (int) n_silo_matset["field_mixvar_values"].dtype().number_of_elements();
-
-        float *mixvals_ptr = NULL;
-        if(n_silo_matset["field_mixvar_values"].dtype().is_float())
-        {
-            mixvals_ptr = n_silo_matset["field_mixvar_values"].as_float_ptr();
-        }
-        else
-        {
-            n_silo_matset["field_mixvar_values"].to_float_array(
-                                        n_silo_matset["field_mixvar_values_float"]);
-            mixvals_ptr = n_silo_matset["field_mixvar_values_float"].as_float_ptr();
-        }
-
-        avtMixedVariable *mvar = new avtMixedVariable(mixvals_ptr,
-                                                      mix_len,
-                                                      abs_varname_str);
-        void_ref_ptr mvar_ref = void_ref_ptr(mvar, avtMixedVariable::Destruct);
-        cache->CacheVoidRef(abs_varname_str.c_str(),
-                            AUXILIARY_DATA_MIXED_VARIABLE,
-                            timestep,
-                            domain,
-                            mvar_ref);
-    }
-
-    return res;
+        std::ostringstream err_oss;
+        err_oss <<  "Conduit Exception in Blueprint Plugin "
+                    << "avtBlueprintFileFormat::GetVar: " << endl
+                    << e.message();
+        BP_PLUGIN_EXCEPTION1(VisItException, err_oss.str());
+    }    
 }
 
 
@@ -1999,6 +2009,8 @@ avtBlueprintFileFormat::GetAuxiliaryData(const char *var,
 //  Creation:   December 8, 2020
 //
 //  Modifications:
+//     Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//     Encased in try-catch block.
 //
 //
 // ****************************************************************************
@@ -2006,85 +2018,96 @@ avtMaterial *
 avtBlueprintFileFormat::GetMaterial(int domain,
                                     const char *mat_name)
 {
-    BP_PLUGIN_INFO("avtBlueprintFileFormat::GetMaterial " 
-                    << domain << " "
-                    << mat_name);
-
-    Node n_matset;
-    ReadBlueprintMatset(domain,
-                        mat_name,
-                        n_matset);
-
-    std::vector<std::string> matnames = n_matset["matnames"].child_names();
-
-    // use to_silo util to convert from bp to the mixslot rep
-    // that silo and visit use
-
-    Node n_silo_matset;
-    conduit::blueprint::mesh::matset::to_silo(n_matset,
-                                              n_silo_matset);
-
-    int nmats = (int) matnames.size();
-    int nzones = (int) n_silo_matset["matlist"].dtype().number_of_elements();
-
-    int *matlist  = NULL;
-    int *mix_mat  = NULL;
-    int *mix_next = NULL;
-
-    // we need int ptrs for the avtMaterial object,
-    // convert if needed
-
-    Node n_tmp;
-    if(!n_silo_matset["matlist"].dtype().is_int())
+    try
     {
-        n_silo_matset["matlist"].to_int_array(n_tmp["matlist"]);
-        n_silo_matset["mix_mat"].to_int_array(n_tmp["mix_mat"]);
-        n_silo_matset["mix_next"].to_int_array(n_tmp["mix_next"]);
-        matlist  = n_tmp["matlist"].as_int_ptr();
-        mix_mat  = n_tmp["mix_mat"].as_int_ptr();
-        mix_next = n_tmp["mix_next"].as_int_ptr();
-    }
-    else
-    {
-        matlist  = n_silo_matset["matlist"].as_int_ptr();
-        mix_mat  = n_silo_matset["mix_mat"].as_int_ptr();
-        mix_next = n_silo_matset["mix_next"].as_int_ptr();
-    }
+        BP_PLUGIN_INFO("avtBlueprintFileFormat::GetMaterial " 
+                        << domain << " "
+                        << mat_name);
 
-    // we need to adjust the matlist.
-    for(int i=0;i<nzones;i++)
-    {
-        if(matlist[i] > 0 )
+        Node n_matset;
+        ReadBlueprintMatset(domain,
+                            mat_name,
+                            n_matset);
+
+        std::vector<std::string> matnames = n_matset["matnames"].child_names();
+
+        // use to_silo util to convert from bp to the mixslot rep
+        // that silo and visit use
+
+        Node n_silo_matset;
+        conduit::blueprint::mesh::matset::to_silo(n_matset,
+                                                  n_silo_matset);
+
+        int nmats = (int) matnames.size();
+        int nzones = (int) n_silo_matset["matlist"].dtype().number_of_elements();
+
+        int *matlist  = NULL;
+        int *mix_mat  = NULL;
+        int *mix_next = NULL;
+
+        // we need int ptrs for the avtMaterial object,
+        // convert if needed
+
+        Node n_tmp;
+        if(!n_silo_matset["matlist"].dtype().is_int())
         {
-            matlist[i]--;
+            n_silo_matset["matlist"].to_int_array(n_tmp["matlist"]);
+            n_silo_matset["mix_mat"].to_int_array(n_tmp["mix_mat"]);
+            n_silo_matset["mix_next"].to_int_array(n_tmp["mix_next"]);
+            matlist  = n_tmp["matlist"].as_int_ptr();
+            mix_mat  = n_tmp["mix_mat"].as_int_ptr();
+            mix_next = n_tmp["mix_next"].as_int_ptr();
         }
+        else
+        {
+            matlist  = n_silo_matset["matlist"].as_int_ptr();
+            mix_mat  = n_silo_matset["mix_mat"].as_int_ptr();
+            mix_next = n_silo_matset["mix_next"].as_int_ptr();
+        }
+
+        // we need to adjust the matlist.
+        for(int i=0;i<nzones;i++)
+        {
+            if(matlist[i] > 0 )
+            {
+                matlist[i]--;
+            }
+        }
+
+        int mix_len  = (int) n_silo_matset["mix_mat"].dtype().number_of_elements();
+
+        float *mix_vf = NULL;
+        if(n_silo_matset["mix_vf"].dtype().is_float())
+        {
+            mix_vf = n_silo_matset["mix_vf"].as_float_ptr();
+        }
+        else
+        {
+            n_silo_matset["mix_vf"].to_float_array(n_silo_matset["mix_vf_float"]);
+            mix_vf = n_silo_matset["mix_vf_float"].as_float_ptr();
+        }
+
+        avtMaterial *mat = new avtMaterial(nmats,    // The number of materials in mats.
+                                           matnames, // material names
+                                           nzones,   // number of zones (len of matlist)
+                                           matlist,  // matlist
+                                           mix_len,  // length of mix arrays
+                                           mix_mat,  // mix_mat array
+                                           mix_next, // mix_next array
+                                           NULL,     // mix_zone array (OPTIONAL)
+                                           mix_vf    // mix_vf array
+                                           );
+
+        return mat;
     }
-
-    int mix_len  = (int) n_silo_matset["mix_mat"].dtype().number_of_elements();
-
-    float *mix_vf = NULL;
-    if(n_silo_matset["mix_vf"].dtype().is_float())
+    catch (conduit::Error &e)
     {
-        mix_vf = n_silo_matset["mix_vf"].as_float_ptr();
+        std::ostringstream err_oss;
+        err_oss <<  "Conduit Exception in Blueprint Plugin "
+                    << "avtBlueprintFileFormat::GetMaterial: " << endl
+                    << e.message();
+        BP_PLUGIN_EXCEPTION1(VisItException, err_oss.str());
     }
-    else
-    {
-        n_silo_matset["mix_vf"].to_float_array(n_silo_matset["mix_vf_float"]);
-        mix_vf = n_silo_matset["mix_vf_float"].as_float_ptr();
-    }
-
-    avtMaterial *mat = new avtMaterial(nmats,    // The number of materials in mats.
-                                       matnames, // material names
-                                       nzones,   // number of zones (len of matlist)
-                                       matlist,  // matlist
-                                       mix_len,  // length of mix arrays
-                                       mix_mat,  // mix_mat array
-                                       mix_next, // mix_next array
-                                       NULL,     // mix_zone array (OPTIONAL)
-                                       mix_vf    // mix_vf array
-                                       );
-
-    return mat;
 }
 
 // ****************************************************************************
