+++ src/databases/MFEM/MFEM.xml
@@ -2,12 +2,14 @@
   <Plugin name="MFEM" type="database" label="MFEM" version="1.0" enabled="true" mdspecificcode="false" engspecificcode="false" onlyengine="false" noengine="false" dbtype="STMD" haswriter="false" hasoptions="false" filePatternsStrict="false" opensWholeDirectory="false">
     <CXXFLAGS>
       ${MFEM_INCLUDE_DIR}
+      ${VISIT_INCLUDE_DIR}/avt/MFEM
       ${VISIT_INCLUDE_DIR}/third_party_builtin
     </CXXFLAGS>
     <LDFLAGS>
       ${MFEM_LIBRARY_DIR}
     </LDFLAGS>
     <LIBS>
+      avtmfem
       ${MFEM_LIB}
     </LIBS>
     <FilePatterns>
