+++ src/databases/MFEM/avtMFEMFileFormat.C
@@ -13,11 +13,6 @@
 #include <string>
 #include <sstream>
 
-#include <vtkFloatArray.h>
-#include <vtkRectilinearGrid.h>
-#include <vtkStructuredGrid.h>
-#include <vtkUnstructuredGrid.h>
-
 #include <avtDatabaseMetaData.h>
 
 #include <DebugStream.h>
@@ -29,23 +24,12 @@
 
 #include <avtResolutionSelection.h>
 
-#include <StringHelpers.h>
+// NOTE: This is from avtmfem lib
+#include "avtMFEMDataAdaptor.h"
+
 #include <visit_gzstream.h>
-#include <vtkUnsignedIntArray.h>
-#include <vtkUnstructuredGrid.h>
-#include <vtkCell.h>
-#include <vtkCellData.h>
-#include <vtkLine.h>
-#include <vtkTriangle.h>
-#include <vtkHexahedron.h>
-#include <vtkQuad.h>
-#include <vtkTetra.h>
-#include <vtkPoints.h>
-#include "mfem.hpp"
 
 #include <JSONRoot.h>
-
-#include <zlib.h>
 
 // Controls whether to create original cell numbers.
 #define CREATE_ORIGINAL_CELL_NUMBERS
@@ -72,6 +56,10 @@ using namespace mfem;
 //
 //  Programmer: harrison37 -- generated by xml2avt
 //  Creation:   Fri May 23 15:16:20 PST 2014
+// 
+//  Modifications:
+//    Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//    Defatult set `m_new_refine` to false. This will change later.
 //
 // ****************************************************************************
 
@@ -80,6 +68,7 @@ avtMFEMFileFormat::avtMFEMFileFormat(const char *filename)
 {
     selectedLOD = 0;
     root        = NULL;
+    m_new_refine = false;
 }
 
 // ****************************************************************************
@@ -524,12 +513,27 @@ avtMFEMFileFormat::GetTime()
 //
 //  Programmer: harrison37 -- generated by xml2avt
 //  Creation:   Fri May 23 15:16:20 PST 2014
+// 
+//  Modifications:
+//     Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//     Take advantage of new avtMFEM lib.
+// 
+//    Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//    `m_new_refine` is passed to `RefineMeshToVTK` so it can choose 
+//    the appropriate MFEM LOR scheme.
 //
 // ****************************************************************************
 vtkDataSet *
 avtMFEMFileFormat::GetMesh(int domain, const char *meshname)
 {
-    return GetRefinedMesh(string(meshname),domain,selectedLOD+1);
+    // get base mesh
+    Mesh *mesh = FetchMesh(string(meshname),domain);
+    vtkDataSet *res_ds = avtMFEMDataAdaptor::RefineMeshToVTK(mesh, 
+                                                             domain, 
+                                                             selectedLOD+1, 
+                                                             m_new_refine);
+    delete mesh;
+    return res_ds;
 }
 
 
@@ -671,131 +675,11 @@ avtMFEMFileFormat::FetchMesh(const std::string &mesh_name,int domain)
 //    Alister Maguire, Thu Jan  2 15:23:13 MST 2020
 //    Casting int to Geom::Type where appropriate. This is required after the
 //    upgrade to mfem 4.0.
+// 
+//    Justin Privitera, Wed Aug 24 10:59:27 PDT 2022
+//    Deleted the function. Use `avtMFEMDataAdaptor::RefineMeshToVTK()`.
 //
 // ****************************************************************************
-vtkDataSet *
-avtMFEMFileFormat::GetRefinedMesh(const std::string &mesh_name, int domain, int lod)
-{
-     // get base mesh
-    Mesh *mesh = FetchMesh(mesh_name,domain);
-     
-    // create output objects
-    vtkUnstructuredGrid *res_ds  = vtkUnstructuredGrid::New(); 
-    vtkPoints           *res_pts = vtkPoints::New();
-   
-    int npts=0;
-    int neles=0;
-    
-    RefinedGeometry *refined_geo;
-    DenseMatrix      pmat;
-   
-    //
-    // find the # of output points and cells at the selected level of 
-    // refinement (we may want to cache this...)
-    //
-    for (int i = 0; i < mesh->GetNE(); i++)
-    {
-        int geom = mesh->GetElementBaseGeometry(i);
-        int ele_nverts = Geometries.GetVertices(geom)->GetNPoints();
-        refined_geo = GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-        npts  += refined_geo->RefPts.GetNPoints();
-        neles += refined_geo->RefGeoms.Size() / ele_nverts;
-    }
-
-    // create the points for the refined topoloy   
-    res_pts->Allocate(npts);
-    res_pts->SetNumberOfPoints((vtkIdType) npts);
-   
-    // create the points for the refined topoloy
-    int pt_idx=0;
-    for (int i = 0; i < mesh->GetNE(); i++)
-    {
-        int geom = mesh->GetElementBaseGeometry(i);
-        refined_geo = GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-        // refined points
-        mesh->GetElementTransformation(i)->Transform(refined_geo->RefPts, pmat);
-        for (int j = 0; j < pmat.Width(); j++)
-        {
-            double pt_vals[3]={0.0,0.0,0.0};
-            pt_vals[0] = pmat(0,j);
-            if (pmat.Height() > 1)
-                pt_vals[1] = pmat(1,j);
-            if (pmat.Height() > 2)
-                pt_vals[2] = pmat(2,j);
-            res_pts->InsertPoint(pt_idx,pt_vals);
-            pt_idx++;
-        }
-    }
-    
-    res_ds->SetPoints(res_pts);
-    res_pts->Delete();  
-    // create the cells for the refined topology   
-    res_ds->Allocate(neles);
-
-#ifdef CREATE_ORIGINAL_CELL_NUMBERS
-    // Make some original cell ids so we can try to eliminate the mesh lines.
-    std::vector<unsigned int> originalCells;
-    originalCells.reserve(neles * 2);
-    unsigned int udomain = static_cast<unsigned int>(domain);
-    bool doOriginalCells = true;
-#endif
-
-    pt_idx=0;
-    for (int i = 0; i <  mesh->GetNE(); i++)
-    {
-        int geom       = mesh->GetElementBaseGeometry(i);
-        int ele_nverts = Geometries.GetVertices(geom)->GetNPoints();
-        refined_geo    = GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-
-        Array<int> &rg_idxs = refined_geo->RefGeoms;
-
-        vtkCell *ele_cell = NULL;
-        // rg_idxs contains all of the verts for the refined elements
-        for (int j = 0; j < rg_idxs.Size(); )
-        {
-            switch (geom)
-            {
-                case Geometry::SEGMENT:
-                    ele_cell = vtkLine::New();
-                    doOriginalCells = false;
-                break;
-                case Geometry::TRIANGLE:     ele_cell = vtkTriangle::New();   break;
-                case Geometry::SQUARE:       ele_cell = vtkQuad::New();       break;
-                case Geometry::TETRAHEDRON:  ele_cell = vtkTetra::New();      break;
-                case Geometry::CUBE:         ele_cell = vtkHexahedron::New(); break;
-            }
-            // the are ele_nverts for each refined element
-            for (int k = 0; k < ele_nverts; k++, j++)
-                ele_cell->GetPointIds()->SetId(k,pt_idx + rg_idxs[j]);
-
-            res_ds->InsertNextCell(ele_cell->GetCellType(),
-                                   ele_cell->GetPointIds());
-            ele_cell->Delete();
-
-#ifdef CREATE_ORIGINAL_CELL_NUMBERS
-            originalCells.push_back(udomain);
-            originalCells.push_back(static_cast<unsigned int>(i));
-#endif
-        }
-
-        pt_idx += refined_geo->RefPts.GetNPoints();
-    }
-
-#ifdef CREATE_ORIGINAL_CELL_NUMBERS
-    vtkUnsignedIntArray *ocn = vtkUnsignedIntArray::New();
-    ocn->SetName("avtOriginalCellNumbers");
-    ocn->SetNumberOfComponents(2);
-    ocn->SetNumberOfTuples(originalCells.size()/2);
-    memcpy(ocn->GetVoidPointer(0), &originalCells[0],
-           sizeof(unsigned int) * originalCells.size());
-    res_ds->GetCellData()->AddArray(ocn);
-    ocn->Delete();
-#endif
-
-    delete mesh;
-       
-    return res_ds;
-}
 
 // ****************************************************************************
 //  Method: avtMFEMFileFormat::GetRefinedVar
@@ -819,6 +703,13 @@ avtMFEMFileFormat::GetRefinedMesh(const std::string &mesh_name, int domain, int 
 //    Alister Maguire, Thu Jan  2 15:23:13 MST 2020
 //    Casting int to Geom::Type where appropriate. This is required after the
 //    upgrade to mfem 4.0.
+// 
+//    Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//    Take advantage of avtMFEM lib.
+// 
+//    Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//    `m_new_refine` is passed to `RefineGridFunctionToVTK` so it can choose 
+//    the appropriate MFEM LOR scheme.
 //
 // ****************************************************************************
 vtkDataArray *
@@ -851,14 +742,24 @@ avtMFEMFileFormat::GetRefinedVar(const std::string &var_name,
         }
     }
 
-    // check for special vars
-    if(var_name == "element_coloring")
-        return GetRefinedElementColoring(mesh_name,domain,lod);
-    else if(var_name == "element_attribute") // handle with materials in the future?
-        return GetRefinedElementAttribute(mesh_name,domain,lod);
+    vtkDataArray *rv;
 
     // get base mesh
     Mesh *mesh = FetchMesh(mesh_name,domain);
+
+    // check for special vars
+    if(var_name == "element_coloring")
+    {
+        rv = avtMFEMDataAdaptor::RefineElementColoringToVTK(mesh, domain, lod);
+        delete mesh;
+        return rv;
+    }
+    else if(var_name == "element_attribute") // handle with materials in the future?
+    {
+        rv = avtMFEMDataAdaptor::RefineElementAttributeToVTK(mesh, lod);
+        delete mesh;
+        return rv;
+    }
 
     JSONRootEntry &field = root->DataSet(mesh_name).Field(var_name);
     string field_path = field.Path().Expand(domain);
@@ -895,68 +796,18 @@ avtMFEMFileFormat::GetRefinedVar(const std::string &var_name,
         gf = new GridFunction(mesh,igf());   
     }
 
-    int npts=0;
-    int neles=0;
-    
-    RefinedGeometry *refined_geo;
-    Vector           scalar_vals;
-    DenseMatrix      vec_vals;
-    DenseMatrix      pmat;
-
-    //
-    // find the # of output points and cells at the selected level of 
-    // refinement (we may want to cache this...)
-    //
-    for (int i = 0; i < mesh->GetNE(); i++)
+    int gf_ncomps = gf->VectorDim();
+    if (gf_ncomps != ncomps)
     {
-        int geom = mesh->GetElementBaseGeometry(i);
-        int ele_nverts = Geometries.GetVertices(geom)->GetNPoints();
-        refined_geo    = GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-        npts  += refined_geo->RefPts.GetNPoints();
-        neles += refined_geo->RefGeoms.Size() / ele_nverts;
+        EXCEPTION1(InvalidVariableException, 
+            "Expected equality of number of components.");
     }
-    
-    vtkFloatArray *rv = vtkFloatArray::New();
-    if(ncomps == 2)
-        rv->SetNumberOfComponents(3);
-    else
-        rv->SetNumberOfComponents(ncomps);
-    if(var_is_nodal)
-        rv->SetNumberOfTuples(npts);
-    else
-        rv->SetNumberOfTuples(neles);
 
-    double tuple_vals[9];
-    int ref_idx=0;
-    for (int i = 0; i <  mesh->GetNE(); i++)
-    {
-        int geom       = mesh->GetElementBaseGeometry(i);
-        refined_geo    = GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-        if(ncomps == 1)
-        {
-            gf->GetValues(i, refined_geo->RefPts, scalar_vals, pmat);
-            for (int j = 0; j < scalar_vals.Size();j++)
-            {
-                tuple_vals[0] = scalar_vals(j);
-                rv->SetTuple(ref_idx, tuple_vals); 
-                ref_idx++;
-            }
-        }
-        else
-        {
-            gf->GetVectorValues(i, refined_geo->RefPts, vec_vals, pmat);
-            for (int j = 0; j < vec_vals.Width(); j++)
-            {
-                tuple_vals[2] = 0.0;
-                tuple_vals[0] = vec_vals(0,j);
-                tuple_vals[1] = vec_vals(1,j);
-                if (vec_vals.Height() > 2)
-                    tuple_vals[2] = vec_vals(2,j);
-                rv->SetTuple(ref_idx, tuple_vals); 
-                ref_idx++;
-            }
-        }
-    }
+    rv = avtMFEMDataAdaptor::RefineGridFunctionToVTK(mesh, 
+                                                     gf, 
+                                                     lod, 
+                                                     m_new_refine, 
+                                                     var_is_nodal);
     
     delete gf;
     delete mesh;
@@ -987,70 +838,12 @@ avtMFEMFileFormat::GetRefinedVar(const std::string &var_name,
 //   Alister Maguire, Thu Jan  2 15:23:13 MST 2020
 //   Casting int to Geom::Type where appropriate. This is required after the
 //   upgrade to mfem 4.0.
+// 
+//    Justin Privitera, Wed Aug 24 10:59:27 PDT 2022
+//    Deleted the function. Use 
+//    `avtMFEMDataAdaptor::RefineElementColoringToVTK()`.
 //
 // ****************************************************************************
-vtkDataArray *
-avtMFEMFileFormat::GetRefinedElementColoring(const std::string &mesh_name,
-                                             int domain,
-                                             int lod)
-{
-    //
-    // fetch the base mfem mesh
-    //
-    Mesh *mesh = FetchMesh(mesh_name,domain);
-    int npts=0;
-    int neles=0;
-    
-    RefinedGeometry *refined_geo;
-    Array<int>       coloring;
-    
-    //
-    // find the # of output points and cells at the selected level of 
-    // refinement (we may want to cache this...)
-    //
-    for (int i = 0; i < mesh->GetNE(); i++)
-    {
-        int geom = mesh->GetElementBaseGeometry(i);
-        int ele_nverts = Geometries.GetVertices(geom)->GetNPoints();
-        refined_geo    = GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-        npts  += refined_geo->RefPts.GetNPoints();
-        neles += refined_geo->RefGeoms.Size() / ele_nverts;
-    }
-
-    vtkFloatArray *rv = vtkFloatArray::New();
-    rv->SetNumberOfComponents(1);
-    rv->SetNumberOfTuples(neles);
-
-    //
-    // Use mfem's mesh coloring alog
-    //
-
-    // seed using domain id for predictable results
-    srand(domain);
-
-    double a = double(rand()) / (double(RAND_MAX) + 1.);
-
-    int el0 = (int)floor(a * mesh->GetNE());
-    mesh->GetElementColoring(coloring, el0);
-    int ref_idx=0;
-    // set output array value from generated coloring
-    for (int i = 0; i < mesh->GetNE(); i++)
-    {
-        int geom = mesh->GetElementBaseGeometry(i);
-        int nv = Geometries.GetVertices(geom)->GetNPoints();
-        refined_geo= GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-        for (int j = 0; j < refined_geo->RefGeoms.Size(); j += nv)
-        {
-             rv->SetTuple1(ref_idx,coloring[i]+1);
-             ref_idx++;
-        }
-   }
-   
-   delete mesh;
-   
-   return rv;
-}
-
 
 // ****************************************************************************
 //  Method: avtMFEMFileFormat::GetRefinedElementAttribute
@@ -1072,60 +865,12 @@ avtMFEMFileFormat::GetRefinedElementColoring(const std::string &mesh_name,
 //    Alister Maguire, Thu Jan  2 15:23:13 MST 2020
 //    Casting int to Geom::Type where appropriate. This is required after the
 //    upgrade to mfem 4.0.
-//
+// 
+//    Justin Privitera, Wed Aug 24 10:59:27 PDT 2022
+//    Deleted the function. Use 
+//    `avtMFEMDataAdaptor::RefineElementAttributeToVTK()`.
+// 
 // ****************************************************************************
-vtkDataArray *
-avtMFEMFileFormat::GetRefinedElementAttribute(const std::string &mesh_name, 
-                                              int domain, 
-                                              int lod)
-{
-    //
-    // fetch the base mfem mesh
-    //
-    Mesh *mesh = FetchMesh(mesh_name,domain);
-    int npts=0;
-    int neles=0;
-    
-    RefinedGeometry *refined_geo;
-    Array<int>       coloring;
-    
-    //
-    // find the # of output points and cells at the selected level of 
-    // refinement (we may want to cache this...)
-    //
-    for (int i = 0; i < mesh->GetNE(); i++)
-    {
-        int geom = mesh->GetElementBaseGeometry(i);
-        int ele_nverts = Geometries.GetVertices(geom)->GetNPoints();
-        refined_geo    = GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-        npts  += refined_geo->RefPts.GetNPoints();
-        neles += refined_geo->RefGeoms.Size() / ele_nverts;
-    }
-
-    vtkFloatArray *rv = vtkFloatArray::New();
-    rv->SetNumberOfComponents(1);
-    rv->SetNumberOfTuples(neles);
-
-    // set output array value from the mfem mesh's "Attribue" field
-    int ref_idx=0;
-    for (int i = 0; i < mesh->GetNE(); i++)
-    {
-        int geom = mesh->GetElementBaseGeometry(i);
-        int nv = Geometries.GetVertices(geom)->GetNPoints();
-        refined_geo= GlobGeometryRefiner.Refine((Geometry::Type)geom, lod, 1);
-        int attr = mesh->GetAttribute(i);
-        for (int j = 0; j < refined_geo->RefGeoms.Size(); j += nv)
-        {
-             rv->SetTuple1(ref_idx,attr);
-             ref_idx++;
-        }
-   }
-   
-   delete mesh;
-   
-   return rv;
-}
-
 
 // ****************************************************************************
 //  Method: avtMFEMFileFormat::RegisterDataSelections
@@ -1154,6 +899,3 @@ avtMFEMFileFormat::RegisterDataSelections(const std::vector<avtDataSelection_p>&
         }
     }
 }
-
-
-
