+++ src/databases/MFEM/avtMFEMFileFormat.h
@@ -37,6 +37,16 @@ class JSONRoot;
 //
 //   Mark C. Miller, Mon Dec 11 15:50:16 PST 2017
 //   Add support for an mfem_cat file
+// 
+//   Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//   Removed `GetRefinedMesh`, `GetRefinedElementColoring`, and
+//   `GetRefinedElementAttribute`. These methods can be found in the avtMFEM
+//   lib.
+// 
+//    Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//    Added `m_new_refine`, a boolean to toggle between using new MFEM LOR or 
+//    legacy MFEM LOR. When true, the new scheme is enabled. It is never set to
+//    true yet, but will be in the future.
 // ****************************************************************************
 
 class avtMFEMFileFormat : public avtSTMDFileFormat
@@ -75,23 +85,11 @@ class avtMFEMFileFormat : public avtSTMDFileFormat
 
     mfem::Mesh                     *FetchMesh(const std::string &mesh_name,
                                               int chunk);
-                                              
-    vtkDataSet                      *GetRefinedMesh(const std::string &mesh_name,
-                                                    int chunk,
-                                                    int lod);
-                                                    
+                                                                                                  
     vtkDataArray                    *GetRefinedVar(const std::string &mesh_name,
                                                    int chunk,
                                                    int lod);
                                                    
-    vtkDataArray                    *GetRefinedElementColoring(const std::string &mesh_name, 
-                                                                int domain, 
-                                                                int lod);
-                                                                
-    vtkDataArray                    *GetRefinedElementAttribute(const std::string &mesh_name, 
-                                                                int domain, 
-                                                                int lod);
-                                             
     JSONRoot                        *root;  
 
     // For handling .mfem_cat files
@@ -100,6 +98,8 @@ class avtMFEMFileFormat : public avtSTMDFileFormat
                                                           const std::string &mesh_path,
                                                           std::istringstream &imeshstr);
     std::map<std::string, std::pair<size_t,size_t> > catFileMap;
+
+    bool                            m_new_refine;
 };
 
 #endif
