+++ src/engine/main/Engine.C
@@ -82,6 +82,10 @@
 #include <avtStructuredDomainBoundaries.h>
 #include <avtExecutionManager.h>
 
+#ifdef HAVE_CONDUIT
+#include <avtConduitBlueprintDataAdaptor.h>
+#endif
+
 #include <visit-config.h>
 #if defined(HAVE_OSMESA) || defined(HAVE_EGL)
 #  include <vtkOffScreenRenderingFactory.h>
@@ -767,6 +771,9 @@ public:
 //
 //   Alok Hota, Tue Feb 23 19:10:32 PST 2016
 //   Add support for OSPRay.
+// 
+//   Justin Privitera, Wed Aug 24 11:08:51 PDT 2022
+//   Call `avtConduitBlueprintDataAdaptor::Initialize();`.
 //
 // ****************************************************************************
 
@@ -783,6 +790,9 @@ Engine::InitializeCompute()
     int setupTimer = visitTimer->StartTimer();
     InitVTK::Initialize();
     InitVTKRendering::Initialize();
+#ifdef HAVE_CONDUIT
+    avtConduitBlueprintDataAdaptor::Initialize();
+#endif
     if (avtCallback::GetSoftwareRendering())
     {
         // Install factory for  VisIt's OffScreen Render Window overrides
